requirements.txt
aiofiles==0.7.0
bson==0.5.10
fastapi==0.65.1
filelock==3.0.12
Jinja2==3.0.0
mongoengine==0.23.1
pylint==2.5.3
pydantic==1.8.2
python-decouple==3.4
python-multipart==0.0.5
requests==2.24.0
shortuuid==1.0.1
typing-extensions==3.10.0.0
uvicorn==0.13.4